add confirm admin email template
