+++ consts/email-actions.js
@@ -6,6 +6,7 @@ const EmailActions = {
   UNLOCK_USER: 'UNLOCK_USER',
   INCOMPLETE_OPERATIONS_CART: 'INCOMPLETE_OPERATIONS_CART',
   SUCCESSFUL_ORDER: 'SUCCESSFUL_ORDER',
+  CONFIRM_ADMIN_EMAIL: 'CONFIRM_ADMIN_EMAIL',
 };
 
 module.exports = { EmailActions };
