+++ email-templates/confirm-admin-email.pug
@@ -1,0 +1,10 @@
+extends template
+block subheaderText
+    p.registrationText Реєстрація Адміністратора
+block content
+    div.contentRegistrationText
+        p.congrats Вітаємо! #[br] Щоб зареєструватись як адміністратор, підтвердіть пошту:
+        button.confirm
+            a(href=frontendUrl + `/confirmation/` + token) Підтвердити e-mail
+        p.link або перейдіть за посиланням нижче:
+        a.alink(href=frontendUrl + `/confirmation/` + token) #{frontendUrl + `/confirmation/` + token}
